requirements.txt
pip~=25.2
wheel~=0.45.1
PySocks~=1.7.1
protobuf~=5.29.5
pytest~=8.4.1
tornado~=6.5.1
altgraph~=0.17.4
packaging~=25.0
pefile~=2023.2.7
setuptools~=80.9.0
Cython~=3.1.4
pyinstaller~=6.15.0
psutil~=7.1.0
pillow~=11.2.1
pywin32-ctypes~=0.2.3
cffi~=1.17.1
pywin32~=310
nltk~=3.9.2
traitlets~=5.14.3
pyinstaller-hooks-contrib~=2025.8
jupyter_core~=5.8.1
matplotlib~=3.8.4
numpy~=1.26.4